% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{list_paramIT_PRZcst}
\alias{list_paramIT_PRZcst}
\title{Fit MCMC of \code{propiconazole} data set with IT model}
\usage{
data(list_paramIT_PRZcst)
}
\description{
Fit MCMC of \code{propiconazole} data set with IT model
}
\keyword{data}
\keyword{set}
