% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{list_paramSD_PRZcst}
\alias{list_paramSD_PRZcst}
\title{Fit MCMC of \code{propiconazole} data set with SD model}
\usage{
data(list_paramSD_PRZcst)
}
\description{
Fit MCMC of \code{propiconazole} data set with SD model
}
\keyword{data}
\keyword{set}
