% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKTDsimulations.R
\name{runIT}
\alias{runIT}
\title{Single IT model}
\usage{
runIT(time, conc, kd, hb, alpha, beta)
}
\arguments{
\item{time}{vector of time}

\item{conc}{vector of concentration}

\item{kd}{parameter of TK model}

\item{hb}{paramter background mortality model}

\item{alpha}{parameter median of IT model}

\item{beta}{parameter shape of IT model}
}
\description{
Single IT model
}
