% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKTDsimulations.R
\name{runIT_MCMC}
\alias{runIT_MCMC}
\title{MCMC IT model}
\usage{
runIT_MCMC(time, conc, kd, hb, alpha, beta)
}
\arguments{
\item{time}{vector of time}

\item{conc}{vector of concentration}

\item{kd}{vector of parameter of TK model}

\item{hb}{vector of background mortality model}

\item{alpha}{vector median of IT model}

\item{beta}{vector shape of IT model}
}
\description{
MCMC IT model
}
