% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicationFactor.R
\name{runMFx}
\alias{runMFx}
\title{Function to compute the optimal multiplication factor for SD model
with binary search using distribution of parameters}
\usage{
runMFx(
  time,
  conc,
  model = NULL,
  parms,
  X = 50,
  time_MFx,
  MFx_range = c(0, 1000),
  accuracy = 0.01,
  threshold_iter = 100,
  quiet = FALSE
)
}
\arguments{
\item{time}{vector of time}

\item{conc}{vector of concentration}

\item{model}{model type SD or IT}

\item{parms}{list of parameters}

\item{X}{percentage of reduction of survival}

\item{time_MFx}{time at which Mulitplication Factor has to be evaluated}

\item{MFx_range}{range of Multiplication factor to investigate}

\item{accuracy}{accuracy for the binary search algorithm}

\item{threshold_iter}{maximum number of iteration}

\item{quiet}{If \code{TRUE}, remove log information}
}
\description{
Function to compute the optimal multiplication factor for SD model
with binary search using distribution of parameters
}
