% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKTDsimulations.R
\name{runSD}
\alias{runSD}
\title{Single SD model}
\usage{
runSD(time, conc, kd, hb, z, kk)
}
\arguments{
\item{time}{vector of time}

\item{conc}{vectopr of concentration}

\item{kd}{parameter of TK model}

\item{hb}{paramter background mortality model}

\item{z}{parameter threshold of SD model}

\item{kk}{parameter killing rate of SD model}
}
\description{
Single SD model
}
