% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKTDsimulations.R
\name{runSD_MCMC}
\alias{runSD_MCMC}
\title{MCMC SD model}
\usage{
runSD_MCMC(time, conc, kd, hb, z, kk)
}
\arguments{
\item{time}{vector of time}

\item{conc}{vectopr of concentration}

\item{kd}{vector of parameter for TK model}

\item{hb}{vector of background mortality model}

\item{z}{vector of threshold of SD model}

\item{kk}{vector of killing rate of SD model}
}
\description{
MCMC SD model
}
