% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKTDsimulations.R
\name{runTK}
\alias{runTK}
\title{Single TK model}
\usage{
runTK(time, conc, kd)
}
\arguments{
\item{time}{vector of time}

\item{conc}{vectopr of concentration}

\item{kd}{parameter of TK model}
}
\description{
Single TK model
}
