% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKTDsimulations.R
\name{runTK_MCMC}
\alias{runTK_MCMC}
\title{MCMC TK model}
\usage{
runTK_MCMC(time, conc, kd)
}
\arguments{
\item{time}{vector of time}

\item{conc}{vectopr of concentration}

\item{kd}{vector of parameter of TK model}
}
\description{
MCMC TK model
}
