% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tktdjl2r.R
\name{tktdjl2r_setup}
\alias{tktdjl2r_setup}
\title{Setup tktdjl2}
\usage{
tktdjl2r_setup(...)
}
\arguments{
\item{...}{Parameters are passed down to JuliaCall::julia_setup}
}
\description{
This function initializes Julia and the TKTDsimulation.jl package.
The first time will be long since it includes precompilation.
}
\examples{

\donttest{ ## tktdjl2r_setup() is time-consuming and requires Julia+TKTDsimulations.jl

tktdjl2r::tktdjl2r_setup()

}

}
